Move wallet code from simple.py to wallet.py, write tests, remove mpl.py .
