+++ decred/decred/wallet/wallet.py
@@ -4,10 +4,15 @@ Copyright (c) 2019-2020, The Decred developers
 See LICENSE for details
 """
 
+from pathlib import Path
+
 from decred import DecredError
 from decred.crypto import crypto, mnemonic, rando
-from decred.util import chains, encode, helpers
-from decred.util.database import KeyValueDatabase
+from decred.dcr import nets
+from decred.dcr.dcrdata import DcrdataBlockchain
+from decred.util import chains, database, encode, helpers
+from decred.util.helpers import mkdir
+
 
 from . import accounts
 
@@ -40,7 +45,7 @@ class Wallet:
         Args:
             path (str): The path to the wallet database.
         """
-        self.db = KeyValueDatabase(path)
+        self.db = database.KeyValueDatabase(path)
         self.masterDB = self.db.child("master", blobber=encode.ByteArray)
         self.coinDB = self.db.child(
             "accts", datatypes=("INTEGER", "BLOB"), blobber=accounts.AccountManager
@@ -106,7 +111,7 @@ class Wallet:
         Args:
             words (list(str)): Mnemonic seed. Assumed to be PGP words.
             path (str): Filepath to store wallet.
-            password (str): User password. Passed to Wallet.create.
+            password (str): User password. Passed to Wallet.initialize.
 
         Returns:
             Wallet: A wallet initialized from the seed parsed from `words`.
@@ -116,7 +121,7 @@ class Wallet:
         userSeed = decoded[:-1]
         cs = crypto.sha256ChecksumByte(userSeed.b)
         if cs != cksum:
-            raise Exception("bad checksum %r != %r" % (cs, cksum))
+            raise DecredError("bad checksum %r != %r" % (cs, cksum))
         wallet = Wallet(path)
         wallet.initialize(userSeed.b, password.encode(), netParams)
         userSeed.zero()
@@ -146,3 +151,95 @@ class Wallet:
         acctMgr.load(acctDB, signals)
         self.mgrCache[coinType] = acctMgr
         return acctMgr
+
+
+# SimpleWallet code.
+
+
+class DefaultSignals:
+    """DefaultSignals prints the balance to stdout."""
+
+    @staticmethod
+    def balance(b):
+        print(repr(b))
+
+
+def paths(base, network):
+    """
+    Get the paths for the network directory, the wallet database file, and the
+    blockchain database file.
+    """
+    netDir = Path(base) / network
+    dbPath = netDir / "wallet.db"
+    dcrPath = netDir / "dcr.db"
+    return netDir, dbPath, dcrPath
+
+
+DCRDATA_PATHS = {
+    "mainnet": "https://explorer.dcrdata.org/",
+    "testnet3": "https://testnet.dcrdata.org/",
+    "simnet": "http://localhost:17779",
+}
+
+
+class SimpleWallet(Wallet):
+    """
+    SimpleWallet is a single-account Decred wallet.
+    """
+
+    def __init__(self, walletDir, pw, network, signals=None, allowCreate=False):
+        """
+        Args:
+            dir (str): A directory for wallet database files.
+            pw (str): The user password.
+            network (str): The network name.
+            signals (Signals): A signal processor.
+        """
+        signals = signals if signals else DefaultSignals
+        netParams = nets.parse(network)
+        netDir, dbPath, dcrPath = paths(walletDir, netParams.Name)
+        if not Path(netDir).exists():
+            mkdir(netDir)
+        dcrdataDB = database.KeyValueDatabase(dcrPath)
+        # The initialized DcrdataBlockchain will not be connected, as that is a
+        # blocking operation. It will be called when the wallet is open.
+        dcrdataURL = DCRDATA_PATHS[netParams.Name]
+        self.dcrdata = DcrdataBlockchain(dcrdataDB, netParams, dcrdataURL)
+        chains.registerChain("dcr", self.dcrdata)
+        walletExists = Path(dbPath).is_file()
+        if not walletExists and not allowCreate:
+            raise DecredError("Wallet does not exist at %s", dbPath)
+
+        super().__init__(dbPath)
+        # words is only set the first time a wallet is created.
+        if not walletExists:
+            seed = rando.newKeyRaw()
+            self.initialize(seed, pw.encode(), netParams)
+            self.words = mnemonic.encode(seed)
+
+        cryptoKey = self.cryptoKey(pw)
+        acctMgr = self.accountManager(chains.BipIDs.decred, signals)
+        self.account = acctMgr.openAccount(0, cryptoKey)
+        self.account.sync()
+
+    def __getattr__(self, name):
+        """Delegate unknown methods to the account."""
+        return getattr(self.account, name)
+
+    @staticmethod
+    def create(walletDir, pw, network, signals=None):
+        """
+        Create a new wallet. Will not overwrite an existing wallet file. All
+        arguments are the same as the SimpleWallet constructor.
+        """
+        netParams = nets.parse(network)
+        _, dbPath, _ = paths(walletDir, netParams.Name)
+        if Path(dbPath).is_file():
+            raise DecredError("wallet already exists at %s" % dbPath)
+        wallet = SimpleWallet(walletDir, pw, network, signals, True)
+        words = wallet.words
+        wallet.words.clear()
+        return wallet, words
+
+    def close(self):
+        self.dcrdata.close()
