+++ decred/examples/create_testnet_wallet.py
@@ -7,7 +7,7 @@ testnet wallet. The mnemonic seed and an address are printed.
 
 from getpass import getpass
 
-from decred.wallet.simple import SimpleWallet
+from decred.wallet.wallet import SimpleWallet
 
 
 def main():
